requirements.txt
flask==2.0.1
praw==7.7.1
asyncpraw==7.7.1
requests==2.26.0
python-dotenv==0.19.0
transformers==4.30.2
torch==2.0.1
numpy==1.24.3
pandas==1.3.5
scikit-learn==0.24.2
newsapi-python==0.2.7
sentence-transformers==2.2.2
tqdm==4.65.0
accelerate==0.27.2
bitsandbytes==0.42.0
einops==0.7.0
safetensors==0.4.2 